refactor: merge *_options.go files into their respective handler files
Each handler package had a separate *_options.go containing only the option
type alias and WithXxx setter functions — trivial boilerplate that added no
meaningful separation. Moved them inline at the bottom of the primary file
(middleware.go / handler.go), reducing file count by 8 without changing any
public API.
