+++ internal/handler/proxy/handler.go
@@ -169,3 +169,29 @@ func (h *Handler) logger(request *http.Request) contracts.Logger {
 
 	return h.proxyLogger
 }
+
+type HandlerOption = func(*Handler)
+
+func WithURLReplacerFactory(replacerFactory urlreplacer.ReplacerFactory) HandlerOption {
+	return func(m *Handler) {
+		m.replacers = replacerFactory
+	}
+}
+
+func WithHTTPClient(http contracts.HTTPClient) HandlerOption {
+	return func(m *Handler) {
+		m.http = http
+	}
+}
+
+func WithProxyLogger(logger contracts.Logger) HandlerOption {
+	return func(m *Handler) {
+		m.proxyLogger = logger
+	}
+}
+
+func WithRewriteLogger(logger contracts.Logger) HandlerOption {
+	return func(m *Handler) {
+		m.rewriteLogger = logger
+	}
+}
