+++ internal/handler/script/handler.go
@@ -68,3 +68,23 @@ func (h *Handler) runScript(luaState *lua.LState) error {
 
 	return luaState.DoString(string(scriptContent))
 }
+
+type HandlerOption = func(*Handler)
+
+func WithLogger(logger contracts.Logger) HandlerOption {
+	return func(h *Handler) {
+		h.logger = logger
+	}
+}
+
+func WithScript(script config.Script) HandlerOption {
+	return func(h *Handler) {
+		h.script = script
+	}
+}
+
+func WithFileSystem(fs afero.Fs) HandlerOption {
+	return func(h *Handler) {
+		h.fs = fs
+	}
+}
