+++ internal/handler/uncors_handler.go
@@ -100,3 +100,59 @@ func replaceWildcards(host string) string {
 
 	return host
 }
+
+type RequestHandlerOption = func(*RequestHandler)
+
+func WithLogger(logger contracts.Logger) RequestHandlerOption {
+	return func(h *RequestHandler) {
+		h.logger = logger
+	}
+}
+
+func WithMappings(mappings config.Mappings) RequestHandlerOption {
+	return func(h *RequestHandler) {
+		h.mappings = mappings
+	}
+}
+
+func WithCacheMiddlewareFactory(factory CacheMiddlewareFactory) RequestHandlerOption {
+	return func(h *RequestHandler) {
+		h.cacheMiddlewareFactory = factory
+	}
+}
+
+func WithProxyHandlerFactory(factory ProxyHandlerFactory) RequestHandlerOption {
+	return func(h *RequestHandler) {
+		h.proxyHandlerFactory = factory
+	}
+}
+
+func WithStaticHandlerFactory(factory StaticMiddlewareFactory) RequestHandlerOption {
+	return func(h *RequestHandler) {
+		h.staticMiddlewareFactory = factory
+	}
+}
+
+func WithMockHandlerFactory(factory MockHandlerFactory) RequestHandlerOption {
+	return func(h *RequestHandler) {
+		h.mockHandlerFactory = factory
+	}
+}
+
+func WithScriptHandlerFactory(factory ScriptHandlerFactory) RequestHandlerOption {
+	return func(h *RequestHandler) {
+		h.scriptHandlerFactory = factory
+	}
+}
+
+func WithRewriteHandlerFactory(factory RewriteMiddlewareFactory) RequestHandlerOption {
+	return func(h *RequestHandler) {
+		h.rewriteMiddlewareFactory = factory
+	}
+}
+
+func WithOptionsHandlerFactory(factory OptionsMiddlewareFactory) RequestHandlerOption {
+	return func(h *RequestHandler) {
+		h.optionsMiddlewareFactory = factory
+	}
+}
